+++ CHANGELOG.md
@@ -4,6 +4,7 @@
   
   * gzip support for stylesheets and javascripts (Christian Lundgren)
   * Extended manifest with info file size (Christian Lundgren)
+  * verbose option
 
 # v 0.4.0